add a test for gjk intersection
- also factor shape sampling into a separate file
